moved geometry generation into its own class
