+++ AstroView2/Geometry/SpaceGeometry.swift
@@ -1,0 +1,13 @@
+//
+//  GeometryFactory.swift
+//  AstroView2
+//
+//  Created by Jeff Doar on 12/20/24.
+//
+
+import Foundation
+import SceneKit
+
+public protocol SpaceGeometry {
+    func createGeometry() throws -> SCNNode
+}
